[UPDATE] Rename ShoppingList.py to views.py to make it more descriptive
