+++ requirements.txt
@@ -3,6 +3,7 @@ certifi==2017.7.27.1
 cffi==1.11.0
 chardet==3.0.4
 click==6.7
+cov-core==1.15.0
 coverage==4.4.1
 coveralls==1.2.0
 docopt==0.6.2
@@ -10,6 +11,7 @@ dominate==2.3.1
 Flask==0.12.2
 Flask-Bootstrap==3.3.7.1
 Flask-Login==0.4.0
+Flask-Script==2.0.6
 Flask-Testing==0.6.2
 Flask-WTF==0.14.2
 gunicorn==19.7.1
@@ -18,6 +20,7 @@ itsdangerous==0.24
 Jinja2==2.9.6
 MarkupSafe==1.0
 nose==1.3.7
+nose-cov==1.6
 pycparser==2.18
 requests==2.18.4
 six==1.11.0
